REFACT : Renaming and moving two files
