+++ src/services/editor_extensions/PropertiesHiding.ts
